salaryinfo02

+++ src/main/java/core/basesyntax/SalaryInfo.java
@@ -1,10 +1,13 @@
 package core.basesyntax;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+
 public class SalaryInfo {
     static final DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("d.MM.yyyy");
-    public static String getSalaryInfo(String[] names, String[] data,
-                                       String dateFrom, String dateTo) {
+
+    public static String getSalaryInfo
+
+            (String[] names, String[] data, String dateFrom, String dateTo) {
         LocalDate from = LocalDate.parse(dateFrom, dateTimeFormatter);
         LocalDate to = LocalDate.parse(dateTo, dateTimeFormatter);
         StringBuilder builder = new StringBuilder("Report for period ");
@@ -16,7 +19,7 @@ public class SalaryInfo {
                 LocalDate workingDate = LocalDate.parse(lineInfo[0], dateTimeFormatter);
                 String employeeName = lineInfo[1];
                 int rate = Integer.parseInt(lineInfo[2]);
-                int money = Integer.parseInt(lineInfo[3]);
+                long money = Integer.parseInt(lineInfo[3]);
                 if (name.equals(employeeName)) {
                     if ((workingDate.isAfter(from) || workingDate.equals(from))
                             && (workingDate.isBefore(to) || workingDate.equals(to))) {
